app-indicator: Move from "secondary-activate" signal to secondary_activate target
To avoid appindicators writers to use the "secondary-activate"
event to perform actions that can't be done via a menu item
we added a new API:

 - app_indicator_set_secondary_activate_target
 - app_indicator_get_secondary_activate_target

With it, an appindicator writer can simply define a new GtkWidget
that will be activated when a secondary activation will occur.
The GtkWidget must be a sensitive and visible inner child of
the appindicator MenuItem.