scroll-event: change parameter to the right type
The scroll-direction parameter is of type uint, but the annotation says that
it's of Gdk.ScrollDirection. Thus, bindings always see 0 for that parameter.

This commit changes the actual type of the parameter instead of the annotation.
This should be API compatible for C while still maintaining the proper type for
bindings.